refactor to labelinspector.ts
